backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
anthropic>=0.43.0
playwright==1.41.0
Pillow>=10.2.0
numpy>=1.26.3
python-dotenv==1.0.0
supabase>=2.3.0
pydantic>=2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24.0,<0.26
jinja2==3.1.3
